# feat(having): HAVING Phase 4.3 - Performance optimization and memory efficiency