… MEMORY_OPTIMIZATION_SUMMARY.md
… internal/dataframe/having_memory_optimization_test.go
… internal/dataframe/having_optimizer_integration_test.go
… internal/dataframe/having_overhead_analysis_test.go
… internal/dataframe/having_performance_test.go
… internal/dataframe/having_regression_test.go
… memory_profile_test.go